[FIX] 본인인증 기능 구현 버그수정
