+++ src/main/java/com/ssg/usms/business/Identification/util/SmsUtil.java
@@ -1,0 +1,40 @@
+package com.ssg.usms.business.Identification.util;
+
+import net.nurigo.sdk.NurigoApp;
+import net.nurigo.sdk.message.model.Message;
+import net.nurigo.sdk.message.request.SingleMessageSendingRequest;
+import net.nurigo.sdk.message.response.SingleMessageSentResponse;
+import net.nurigo.sdk.message.service.DefaultMessageService;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+
+@Component
+public class SmsUtil {
+
+    @Value("${coolsms.api.key}")
+    private String apiKey;
+    @Value("${coolsms.api.secret}")
+    private String apiSecretKey;
+    private DefaultMessageService messageService;
+
+    @PostConstruct
+    private void init(){
+
+        this.messageService = NurigoApp.INSTANCE.initialize(apiKey,apiSecretKey,"https://api.coolsms.co.kr" );
+    }
+
+    public SingleMessageSentResponse sendOne(String to, String verificationCode) {
+
+        Message message = new Message();
+
+        message.setFrom("010-4046-7715");
+        message.setTo(to);
+        message.setText("[Usms] 아래의 인증번호를 입력해주세요\n" + verificationCode);
+
+        SingleMessageSentResponse response = this.messageService.sendOne(new SingleMessageSendingRequest(message));
+        return response;
+    }
+
+}
