--- src/main/java/com/ssg/usms/business/Security/login/UsmsLoginProvider.java
@@ -1,51 +1,0 @@
-package com.ssg.usms.business.Security.login;
-
-import com.ssg.usms.business.Security.login.service.UsmsUserDetailsService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.*;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Component;
-
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class UsmsLoginProvider implements AuthenticationProvider {
-
-    private final UsmsUserDetailsService usmsUserDetailsService;
-    private final BCryptPasswordEncoder passwordEncoder;
-
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-
-        String username = authentication.getName();
-        String password = (String) authentication.getCredentials();
-
-        UserDetails userDetails = usmsUserDetailsService.loadUserByUsername(username);
-
-        if(userDetails == null || !passwordEncoder.matches(password, userDetails.getPassword())){
-
-            throw new BadCredentialsException("아이디 또는 비밀번호가 일치하지 않습니다.");
-        }
-        if(!userDetails.isAccountNonExpired() || !userDetails.isAccountNonExpired()){
-
-            throw new AccountExpiredException("해당 계정은 만료되었습니다.");
-        }
-        if(!userDetails.isEnabled()){
-
-            throw new DisabledException("해당 계정은 사용이 불가합니다.");
-        }
-        return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-
-        return authentication.equals(UsernamePasswordAuthenticationToken.class);
-    }
-}
