--- src/main/java/com/ssg/usms/business/Security/login/persistence/ResponseLoginDto.java
@@ -1,20 +1,0 @@
-package com.ssg.usms.business.Security.login.persistence;
-
-
-import com.ssg.usms.business.Security.login.UsmsUserDetails;
-import com.ssg.usms.business.user.repository.UsmsUser;
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-
-public class ResponseLoginDto {
-
-    private int code;
-
-    private String message;
-
-    private UsmsUserDetails user;
-}
