+++ src/main/java/com/ssg/usms/business/security/login/authority/UsmsAccessDeniedHandler.java
@@ -1,0 +1,35 @@
+package com.ssg.usms.business.security.login.authority;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ssg.usms.business.error.ErrorResponseDto;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+public class UsmsAccessDeniedHandler implements AccessDeniedHandler {
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
+
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+
+        ErrorResponseDto dto = ErrorResponseDto.builder()
+                .code(HttpStatus.FORBIDDEN.value())
+                .message(accessDeniedException.getMessage())
+                .build();
+
+        response.setStatus(HttpStatus.FORBIDDEN.value());
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.getWriter().write(objectMapper.writeValueAsString(dto));
+
+    }
+
+}
