+++ src/main/java/com/ssg/usms/business/security/login/persistence/RequestLoginDto.java
@@ -1,0 +1,13 @@
+package com.ssg.usms.business.security.login.persistence;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class RequestLoginDto {
+
+    private String username;
+
+    private String password;
+}
